Fixed ci build (#1852)
