+++ packages/components/package.json
@@ -33,6 +33,7 @@
     "@fortawesome/react-fontawesome": "^0.2.0",
     "@react-spectrum/theme-default": "^3.5.1",
     "@react-types/shared": "^3.22.1",
+    "@react-types/textfield": "^3.9.1",
     "bootstrap": "4.6.2",
     "classnames": "^2.3.1",
     "event-target-shim": "^6.0.2",
