+++ packages/components/src/spectrum/collections.ts
@@ -1,0 +1,12 @@
+export {
+  ActionBar,
+  type SpectrumActionBarProps as ActionBarProps,
+  ActionMenu,
+  type SpectrumActionMenuProps as ActionMenuProps,
+  ListView,
+  type SpectrumListViewProps as ListViewProps,
+  MenuTrigger,
+  type SpectrumMenuTriggerProps as MenuTriggerProps,
+  TagGroup,
+  type SpectrumTagGroupProps as TagGroupProps,
+} from '@adobe/react-spectrum';
