+++ packages/components/src/spectrum/content.ts
@@ -1,0 +1,24 @@
+export {
+  Avatar,
+  type SpectrumAvatarProps as AvatarProps,
+  Content,
+  type ContentProps,
+  Divider,
+  type SpectrumDividerProps as DividerProps,
+  Footer,
+  type FooterProps,
+  Heading,
+  type HeadingProps,
+  IllustratedMessage,
+  type SpectrumIllustratedMessageProps as IllustratedMessageProps,
+  Image,
+  type SpectrumImageProps as ImageProps,
+  Keyboard,
+  type KeyboardProps,
+  Text,
+  type TextProps,
+  View,
+  type ViewProps,
+  Well,
+  type SpectrumWellProps as WellProps,
+} from '@adobe/react-spectrum';
