+++ packages/components/src/spectrum/forms.ts
@@ -1,0 +1,19 @@
+export {
+  CheckboxGroup,
+  type SpectrumCheckboxGroupProps as CheckboxGroupProps,
+  Form,
+  type SpectrumFormProps as FormProps,
+  NumberField,
+  type SpectrumNumberFieldProps as NumberFieldProps,
+  RangeSlider,
+  type SpectrumRangeSliderProps as RangeSliderProps,
+  Slider,
+  type SpectrumSliderProps as SliderProps,
+  Switch,
+  type SpectrumSwitchProps as SwitchProps,
+  TextArea,
+  TextField,
+  type SpectrumTextFieldProps as TextFieldProps,
+} from '@adobe/react-spectrum';
+
+export { type SpectrumTextAreaProps as TextAreaProps } from '@react-types/textfield';
