+++ packages/components/src/spectrum/layout.ts
@@ -1,0 +1,6 @@
+export {
+  Flex,
+  type FlexProps,
+  Grid,
+  type GridProps,
+} from '@adobe/react-spectrum';
