+++ packages/components/src/spectrum/pickers.ts
@@ -1,0 +1,4 @@
+export {
+  ComboBox,
+  type SpectrumComboBoxProps as ComboBoxProps,
+} from '@adobe/react-spectrum';
