+++ packages/components/src/spectrum/status.ts
@@ -1,0 +1,16 @@
+export {
+  Badge,
+  type SpectrumBadgeProps as BadgeProps,
+  InlineAlert,
+  type SpectrumInlineAlertProps as InlineAlertProps,
+  LabeledValue,
+  type SpectrumLabeledValueProps as LabeledValueProps,
+  Meter,
+  type SpectrumMeterProps as MeterProps,
+  ProgressBar,
+  type SpectrumProgressBarProps as ProgressBarProps,
+  ProgressCircle,
+  type SpectrumProgressCircleProps as ProgressCircleProps,
+  StatusLight,
+  type SpectrumStatusLightProps as StatusLightProps,
+} from '@adobe/react-spectrum';
